refactor(governance): rename policyEngine to policyEvaluator
- Rename policyEngine.ts to policyEvaluator.ts for improved naming clarity
- Update import in gate.ts to reference policyEvaluator module
- Update export in index.ts to reference policyEvaluator module
- Better aligns module name with its responsibility of evaluating policies
