+++ packages/components/src/governance/index.ts
@@ -1,6 +1,6 @@
 export * from './types'
 export * from './policyLoader'
-export * from './policyEngine'
+export * from './policyEvaluator'
 export * from './auditLogger'
 export * from './gate'
 export * from './governedTool'
